Enabled http-caching on jobs endpoints
